Uploading files works. Files are uploaded to the COM3014-source/uploads directory with the file format <timestamp>-<userID>.jpg. Only image files less than 5MB in size can be uploaded. Images can be queried by accessing /getMoreImages/X.htm, where X must be an index starting at zero and incrementing by one each time it is called to get the next 6 images. Technically this is a RESTful service, so this counts for the marks... right?
Also, removed a bunch more of the hello world crap, along with the test newjsp.jsp that was included in my last commit, and formatted the JSON jsps to look nicer.

Also attempted to make sessions work. They don't. But there's some abstraction that's come out of it which is nice.